Build a Complete nflfastR Data Set

build_nflfastR_pbp is a convenient wrapper around 6 nflfastR functions:

fast_scraper()

clean_pbp()

add_qb_epa()

add_xyac()

add_xpass()

decode_player_ids()

Please see either the documentation of each function or the nflfastR Field Descriptions website to learn about the output.

Usage

build_nflfastR_pbp(game_ids, ..., decode = TRUE, rules = TRUE)

Arguments

game_ids

Vector of character ids or a data frame including the variable game_id (see details for further information).

...

Additional arguments passed to the scraping functions (for internal use)

decode

If TRUE, the function decode_player_ids() will be executed.

rules

If FALSE, printing of the header and footer in the console output will be suppressed.

Value

An nflfastR play-by-play data frame like it can be loaded from https://github.com/nflverse/nflverse-data.

Details

To load valid game_ids please use the package function fast_scraper_schedules().

See also

Examples

# \donttest{
# Build nflfastR pbp for the 2018 and 2019 Super Bowls
build_nflfastR_pbp(c("2018_21_NE_LA", "2019_21_SF_KC"))
#> ── Build nflfastR Play-by-Play Data ───────────── nflfastR version 4.4.0.9005 ──
#> • 12:26:44 | Start download of 2 games...
#> ℹ It is recommended to use parallel processing when trying to load multiple games.Please consider running `future::plan("multisession")`! Will go on sequentially...
#> ✔ 12:26:49 | Download finished. Adding variables...
#> ✔ 12:26:49 | added game variables
#> ✔ 12:26:49 | added nflscrapR variables
#> ✔ 12:26:50 | added ep variables
#> ✔ 12:26:50 | added air_yac_ep variables
#> ✔ 12:26:51 | added wp variables
#> ✔ 12:26:51 | added air_yac_wp variables
#> ✔ 12:26:51 | added cp and cpoe
#> ✔ 12:26:51 | added fixed drive variables
#> ✔ 12:26:51 | added series variables
#> • 12:26:51 | Cleaning up play-by-play...
#> ✔ 12:26:51 | Cleaning completed
#> ✔ 12:26:51 | added qb_epa
#> • 12:26:51 | Computing xyac...
#> ✔ 12:26:54 | added xyac variables
#> • 12:26:54 | Computing xpass...
#> ✔ 12:26:55 | added xpass and pass_oe
#> ✔ 12:26:55 | Decoding of player ids completed
#> ── DONE ────────────────────────────────────────────────────────────────────────
#> # A tibble: 353 × 372
#>    play_id game_id old_g…¹ home_…² away_…³ seaso…⁴  week posteam poste…⁵ defteam
#>      <dbl> <chr>   <chr>   <chr>   <chr>   <chr>   <int> <chr>   <chr>   <chr>  
#>  1       1 2018_2… 201902… LA      NE      POST       21 NA      NA      NA     
#>  2      38 2018_2… 201902… LA      NE      POST       21 NE      away    LA     
#>  3      67 2018_2… 201902… LA      NE      POST       21 NE      away    LA     
#>  4      89 2018_2… 201902… LA      NE      POST       21 NE      away    LA     
#>  5     111 2018_2… 201902… LA      NE      POST       21 NE      away    LA     
#>  6     133 2018_2… 201902… LA      NE      POST       21 NE      away    LA     
#>  7     155 2018_2… 201902… LA      NE      POST       21 NE      away    LA     
#>  8     182 2018_2… 201902… LA      NE      POST       21 LA      home    NE     
#>  9     204 2018_2… 201902… LA      NE      POST       21 LA      home    NE     
#> 10     226 2018_2… 201902… LA      NE      POST       21 LA      home    NE     
#> # … with 343 more rows, 362 more variables: side_of_field <chr>,
#> #   yardline_100 <dbl>, game_date <chr>, quarter_seconds_remaining <dbl>,
#> #   half_seconds_remaining <dbl>, game_seconds_remaining <dbl>,
#> #   game_half <chr>, quarter_end <dbl>, drive <dbl>, sp <dbl>, qtr <dbl>,
#> #   down <dbl>, goal_to_go <dbl>, time <chr>, yrdln <chr>, ydstogo <dbl>,
#> #   ydsnet <dbl>, desc <chr>, play_type <chr>, yards_gained <dbl>,
#> #   shotgun <dbl>, no_huddle <dbl>, qb_dropback <dbl>, qb_kneel <dbl>, …

# It is also possible to directly use the
# output of `fast_scraper_schedules` as input
library(dplyr, warn.conflicts = FALSE)
fast_scraper_schedules(2020) %>%
  slice_tail(n = 3) %>%
  build_nflfastR_pbp()
#> ── Build nflfastR Play-by-Play Data ───────────── nflfastR version 4.4.0.9005 ──
#> • 12:26:55 | Start download of 3 games...
#> ℹ It is recommended to use parallel processing when trying to load multiple games.Please consider running `future::plan("multisession")`! Will go on sequentially...
#> ✔ 12:27:02 | Download finished. Adding variables...
#> ✔ 12:27:02 | added game variables
#> ✔ 12:27:02 | added nflscrapR variables
#> ✔ 12:27:02 | added ep variables
#> ✔ 12:27:03 | added air_yac_ep variables
#> ✔ 12:27:03 | added wp variables
#> ✔ 12:27:03 | added air_yac_wp variables
#> ✔ 12:27:03 | added cp and cpoe
#> ✔ 12:27:03 | added fixed drive variables
#> ✔ 12:27:03 | added series variables
#> • 12:27:03 | Cleaning up play-by-play...
#> ✔ 12:27:04 | Cleaning completed
#> ✔ 12:27:04 | added qb_epa
#> • 12:27:04 | Computing xyac...
#> ✔ 12:27:05 | added xyac variables
#> • 12:27:05 | Computing xpass...
#> ✔ 12:27:05 | added xpass and pass_oe
#> ✔ 12:27:05 | Decoding of player ids completed
#> ── DONE ────────────────────────────────────────────────────────────────────────
#> ── nflverse play by play ───────────────────────────────────────────────────────
#> ℹ Data updated: 2022-09-12 12:27:05 UTC
#> # A tibble: 545 × 372
#>    play_id game_id old_g…¹ home_…² away_…³ seaso…⁴  week posteam poste…⁵ defteam
#>      <dbl> <chr>   <chr>   <chr>   <chr>   <chr>   <int> <chr>   <chr>   <chr>  
#>  1       1 2020_2… 202101… KC      BUF     POST       20 NA      NA      NA     
#>  2      42 2020_2… 202101… KC      BUF     POST       20 BUF     away    KC     
#>  3      57 2020_2… 202101… KC      BUF     POST       20 BUF     away    KC     
#>  4      78 2020_2… 202101… KC      BUF     POST       20 BUF     away    KC     
#>  5     102 2020_2… 202101… KC      BUF     POST       20 BUF     away    KC     
#>  6     123 2020_2… 202101… KC      BUF     POST       20 BUF     away    KC     
#>  7     145 2020_2… 202101… KC      BUF     POST       20 BUF     away    KC     
#>  8     174 2020_2… 202101… KC      BUF     POST       20 BUF     away    KC     
#>  9     207 2020_2… 202101… KC      BUF     POST       20 BUF     away    KC     
#> 10     236 2020_2… 202101… KC      BUF     POST       20 BUF     away    KC     
#> # … with 535 more rows, 362 more variables: side_of_field <chr>,
#> #   yardline_100 <dbl>, game_date <chr>, quarter_seconds_remaining <dbl>,
#> #   half_seconds_remaining <dbl>, game_seconds_remaining <dbl>,
#> #   game_half <chr>, quarter_end <dbl>, drive <dbl>, sp <dbl>, qtr <dbl>,
#> #   down <dbl>, goal_to_go <dbl>, time <chr>, yrdln <chr>, ydstogo <dbl>,
#> #   ydsnet <dbl>, desc <chr>, play_type <chr>, yards_gained <dbl>,
#> #   shotgun <dbl>, no_huddle <dbl>, qb_dropback <dbl>, qb_kneel <dbl>, …

# \dontshow{
# Close open connections for R CMD Check
future::plan("sequential")
# }
# }