Get a Situation Report on System, nflverse Package Versions and Dependencies

This function gives a quick overview of the versions of R and the operating system as well as the versions of nflverse packages, options, and their dependencies. It's primarily designed to help you get a quick idea of what's going on when you're helping someone else debug a problem.

Usage

report(...)

Arguments

...

Arguments passed on to nflreadr::nflverse_sitrep

pkg

a character vector naming installed packages, or NULL (the default) meaning all nflverse packages. The function checks internally if all packages are installed and informs if that is not the case.

recursive

a logical indicating whether dependencies of pkg and their dependencies (and so on) should be included. Can also be a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"). Character string "all" is shorthand for that vector, character string "most" for the same vector without "Enhances", character string "strong" (default) for the first three elements of that vector.

redact_path

a logical indicating whether options that contain "path" in the name should be redacted, default = TRUE

Details

See nflreadr::nflverse_sitrep for details.

Examples

# \donttest{

report(recursive = FALSE)
#> ── System Info ─────────────────────────────────────────────────────────────────
#> • R version 4.4.1 (2024-06-14) • Running under: Ubuntu 22.04.4 LTS
#> ── Package Status ──────────────────────────────────────────────────────────────
#>    package  installed  cran        dev behind
#> 1 nflfastR 4.6.1.9010 4.6.1 4.6.1.9010       
#> 2 nflplotR      1.3.1 1.3.1      1.3.1       
#> 3 nflreadr   1.4.0.12 1.4.0   1.4.0.12       
#> 4 nflseedR      1.2.0 1.2.0 1.2.0.9001    dev
#> ── Package Options ─────────────────────────────────────────────────────────────
#> • No options set for above packages
#> ── Not Installed ───────────────────────────────────────────────────────────────
#> • nfl4th   ()
#> • nflverse ()
#> ────────────────────────────────────────────────────────────────────────────────
nflverse_sitrep(pkg = "nflreadr", recursive = TRUE)
#> ── System Info ─────────────────────────────────────────────────────────────────
#> • R version 4.4.1 (2024-06-14) • Running under: Ubuntu 22.04.4 LTS
#> ── Package Status ──────────────────────────────────────────────────────────────
#>    package installed  cran      dev behind
#> 1 nflreadr  1.4.0.12 1.4.0 1.4.0.12       
#> ── Package Options ─────────────────────────────────────────────────────────────
#> • No options set for above packages
#> ── Package Dependencies ────────────────────────────────────────────────────────
#> • cachem     (1.1.0)   • glue      (1.7.0)  • graphics (4.4.1)  
#> • cli        (3.6.3)   • memoise   (2.0.1)  • methods  (4.4.1)  
#> • curl       (5.2.1)   • rappdirs  (0.3.3)  • stats    (4.4.1)  
#> • data.table (1.15.4)  • rlang     (1.1.4)  • tools    (4.4.1)  
#> • fastmap    (1.2.0)   • grDevices (4.4.1)  • utils    (4.4.1)  
#> ────────────────────────────────────────────────────────────────────────────────

# }