Compute expected points

for provided plays. Returns the data with probabilities of each scoring event and EP added. The following columns must be present: season, home_team, posteam, roof (coded as 'open', 'closed', or 'retractable'), half_seconds_remaining, yardline_100, ydstogo, posteam_timeouts_remaining, defteam_timeouts_remaining

calculate_expected_points(pbp_data)

Arguments

pbp_data
Play-by-play dataset to estimate expected points for.

Value

The original pbp_data with the following columns appended to it:

ep

expected points.

no_score_prob

probability of no more scoring this half.

opp_fg_prob

probability next score opponent field goal this half.

opp_safety_prob

probability next score opponent safety this half.

opp_td_prob

probability of next score opponent touchdown this half.

fg_prob

probability next score field goal this half.

safety_prob

probability next score safety this half.

td_prob

probability text score touchdown this half.

Details

Computes expected points for provided plays. Returns the data with probabilities of each scoring event and EP added. The following columns must be present:

season

home_team

posteam

roof (coded as 'outdoors', 'dome', or 'open' / 'closed' / NA (retractable))

half_seconds_remaining

yardline_100

down

ydstogo

posteam_timeouts_remaining

defteam_timeouts_remaining