Get team rosters for multiple seasons

Given years return a dataset with each player listed as part of the roster.

fast_scraper_roster(seasons, pp = lifecycle::deprecated())

Arguments

seasons
A vector of 4-digit years associated with given NFL seasons

pp
has no effect and will be removed in a future release.

Value

Data frame where each individual row represents a player in the roster of the given team and season containing the following information:

season

4 digit season year.

team

Team abbreviation.

position

Abbreviation of the player's position (e.g. "QB", "WR", "RB", "CB"...).

depth_chart_position

Starting with the 2020 season: the abbreviation of the players depth_chart_position.

jersey_number

The player's 2 digit jersey number.

status

String indicating the status of the player (e.g. "Active", "Inactive", "Injured Reserve"...) at the update time update_dt (see below)

full_name

Full name of the player.

first_name

First name of the player.

last_name

Last name of the player.

birth_date

Birth date of the player.

height

Height of the player.

weight

Weight of the player.

college

Name of the college the player has attended.

high_school

Name of the High School the player has attended (only non-NA for players who were listed in the 2020 season).

gsis_id

The player's NFL GSIS ID, which can be used to link the player to play-by-play data.

espn_id

The player's ESPN ID (only non-NA for players who were listed in the 2020 season).

sportradar_id

The player's Sportradar ID (only non-NA for players who were listed in the 2020 season).

yahoo_id

The player's Yahoo Sports ID (only non-NA for players who were listed in the 2020 season).

rotowire_id

The player's Rotowire ID (only non-NA for players who were listed in the 2020 season).

update_dt

Date and time when the current entry was last updated (starting with the 2020 season).

headshot_url

URL to a player image (starting in the 2020 season on ESPN servers).

Details

The roster data is accessed via the free to use Sleeper API.

See also

Examples

# \donttest{
# Roster of the 2019 and 2020 seasons
fast_scraper_roster(2019:2020)

#> ℹ It is recommended to use parallel processing when trying to load multiple seasons.
#>   Please consider running `future::plan("multisession")`!
#>   Will go on sequentially...

#> # A tibble: 7,735 x 24
#>    season team  position depth_chart_pos… jersey_number status full_name
#>     <dbl> <chr> <chr>    <chr>                    <int> <chr>  <chr>    
#>  1   2019 ARI   C        NA                          60 CUT    Jacob Oh…
#>  2   2019 ARI   C        NA                          53 ACT    A.Q. Shi…
#>  3   2019 ARI   C        NA                          65 ACT    Lamont G…
#>  4   2019 ARI   CB       NA                          27 SUS    Josh Shaw
#>  5   2019 ARI   CB       NA                          30 ACT    Jalen Da…
#>  6   2019 ARI   CB       NA                          33 ACT    Byron Mu…
#>  7   2019 ARI   CB       NA                          22 CUT    Trevor W…
#>  8   2019 ARI   CB       NA                          26 RES    Brandon …
#>  9   2019 ARI   CB       NA                          NA DEV    Sojourn …
#> 10   2019 ARI   CB       NA                          28 CUT    Deatrick…
#> # … with 7,725 more rows, and 17 more variables: first_name <chr>,
#> #   last_name <chr>, birth_date <date>, height <chr>, weight <chr>,
#> #   college <chr>, high_school <chr>, gsis_id <chr>, espn_id <int>,
#> #   sportradar_id <chr>, yahoo_id <int>, rotowire_id <int>, pff_id <int>,
#> #   update_dt <dttm>, headshot_url <glue>, fantasy_data_id <int>,
#> #   sleeper_id <chr>

# \dontshow{
# Close open connections for R CMD Check
future::plan("sequential")
# }
# }