Load Player Level Weekly Stats

Load Player Level Weekly Stats

Usage

load_player_stats(...)

Arguments

...

Arguments passed on to nflreadr::load_player_stats

seasons

a numeric vector of seasons to return, defaults to most recent season. If set to TRUE, returns all available data.

stat_type

one of "offense", "defense", or "kicking"

file_type

One of c("rds", "qs", "csv", "parquet"). Can also be set globally with options(nflreadr.prefer)

Value

A tibble of week-level player statistics that aims to match NFL official box scores.

See also

Examples

# \donttest{
try({# to avoid CRAN test problems
stats <- load_player_stats()
dplyr::glimpse(stats)
})
#> Warning: We have changed the behavior of `load_player_stats()` as of nflfastR 4.3.0.
#> Calling it without an argument will return the current season only instead of
#> all available seasons.
#> Please try `load_player_stats(seasons = TRUE)` to get all seasons.
#> This warning is displayed once every 8 hours.
#> Rows: 619
#> Columns: 53
#> $ player_id                   <chr> "00-0023459", "00-0023459", "00-0026498", …
#> $ player_name                 <chr> "A.Rodgers", "A.Rodgers", "M.Stafford", "M…
#> $ player_display_name         <chr> "Aaron Rodgers", "Aaron Rodgers", "Matthew…
#> $ position                    <chr> "QB", "QB", "QB", "QB", "QB", "QB", "QB", …
#> $ position_group              <chr> "QB", "QB", "QB", "QB", "QB", "QB", "QB", …
#> $ headshot_url                <chr> "https://static.www.nfl.com/image/upload/f…
#> $ recent_team                 <chr> "NYJ", "NYJ", "LA", "LA", "CAR", "NYJ", "A…
#> $ season                      <int> 2024, 2024, 2024, 2024, 2024, 2024, 2024, …
#> $ week                        <int> 1, 2, 1, 2, 1, 1, 1, 2, 1, 2, 1, 2, 1, 2, …
#> $ season_type                 <chr> "REG", "REG", "REG", "REG", "REG", "REG", …
#> $ opponent_team               <chr> "SF", "TEN", "DET", "ARI", "NO", "SF", "PI…
#> $ completions                 <int> 13, 18, 34, 19, 0, 6, 16, 20, 0, 0, 0, 0, …
#> $ attempts                    <int> 21, 30, 49, 27, 1, 8, 26, 29, 0, 0, 0, 0, …
#> $ passing_yards               <dbl> 167, 176, 317, 216, 0, 36, 155, 241, 0, 0,…
#> $ passing_tds                 <int> 1, 2, 1, 0, 0, 1, 1, 2, 0, 0, 0, 0, 0, 0, …
#> $ interceptions               <dbl> 1, 0, 1, 0, 0, 0, 2, 0, 0, 0, 0, 0, 0, 0, …
#> $ sacks                       <dbl> 1, 2, 2, 5, 0, 0, 2, 1, 0, 0, 0, 0, 0, 0, …
#> $ sack_yards                  <dbl> 5, 12, 13, 24, 0, 0, 18, 8, 0, 0, 0, 0, 0,…
#> $ sack_fumbles                <int> 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ sack_fumbles_lost           <int> 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ passing_air_yards           <dbl> 164, 167, 290, 159, 6, 31, 140, 235, 0, 0,…
#> $ passing_yards_after_catch   <dbl> 93, 96, 171, 104, 0, 24, 73, 101, 0, 0, 0,…
#> $ passing_first_downs         <dbl> 8, 9, 15, 9, 0, 3, 8, 11, 0, 0, 0, 0, 0, 0…
#> $ passing_epa                 <dbl> 3.2582828, 1.6537948, 4.5364209, -11.15959…
#> $ passing_2pt_conversions     <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ pacr                        <dbl> 1.0182927, 1.0538922, 1.0931034, 1.3584906…
#> $ dakota                      <dbl> 0.06657428, 0.04078048, 0.10613713, 0.0396…
#> $ carries                     <int> 1, 2, 0, 0, 0, 1, 1, 4, 0, 0, 0, 0, 0, 0, …
#> $ rushing_yards               <dbl> -1, 1, 0, 0, 0, 7, 0, -1, 0, 0, 0, 0, 0, 0…
#> $ rushing_tds                 <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ rushing_fumbles             <dbl> 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, …
#> $ rushing_fumbles_lost        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ rushing_first_downs         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ rushing_epa                 <dbl> 0.0000000, -0.3271411, NA, NA, NA, 0.77277…
#> $ rushing_2pt_conversions     <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ receptions                  <int> 0, 0, 0, 0, 0, 0, 0, 0, 2, 3, 3, 2, 3, 4, …
#> $ targets                     <int> 0, 0, 0, 0, 0, 0, 0, 0, 3, 5, 4, 3, 4, 4, …
#> $ receiving_yards             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 40, 19, 49, 20, 28…
#> $ receiving_tds               <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ receiving_fumbles           <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ receiving_fumbles_lost      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ receiving_air_yards         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 35, 26, 64, 18, 28…
#> $ receiving_yards_after_catch <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 10, 1, 18, 2, 7, 1…
#> $ receiving_first_downs       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 2, 1, 1, 3, …
#> $ receiving_epa               <dbl> NA, NA, NA, NA, NA, NA, NA, NA, 2.15372744…
#> $ receiving_2pt_conversions   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ racr                        <dbl> NA, NA, NA, NA, NA, NA, NA, NA, 1.1428571,…
#> $ target_share                <dbl> NA, NA, NA, NA, NA, NA, NA, NA, 0.10344828…
#> $ air_yards_share             <dbl> NA, NA, NA, NA, NA, NA, NA, NA, 0.12411348…
#> $ wopr                        <dbl> NA, NA, NA, NA, NA, NA, NA, NA, 0.24205185…
#> $ special_teams_tds           <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ fantasy_points              <dbl> 8.58, 15.14, 14.68, 6.64, 0.00, 6.14, 6.20…
#> $ fantasy_points_ppr          <dbl> 8.58, 15.14, 14.68, 6.64, 0.00, 6.14, 6.20…
# }