Load Official Game Stats

Loads weekly stats for all passers, rushers and receivers in the nflfastR play-by-play data from the 1999 season to the most recent season

load_player_stats(qs = FALSE)

Arguments

qs
Whether to use the function qs::qdeserialize() for more efficient loading.

Value

Weekly stats for all passers, rushers and receivers in the nflfastR play-by-play data from the 1999 season to the most recent season

See also

Examples

# \donttest{
stats <- load_player_stats()
dplyr::glimpse(stats)

#> Rows: 112,757
#> Columns: 38
#> $ player_id                   <chr> "00-0000003", "00-0000003", "00-0000003", …
#> $ player_name                 <chr> "A.al-Jabbar", "A.al-Jabbar", "A.al-Jabbar…
#> $ recent_team                 <chr> "MIA", "MIA", "MIA", "CLE", "CLE", "CLE", …
#> $ season                      <int> 1999, 1999, 1999, 1999, 1999, 1999, 1999, …
#> $ week                        <int> 1, 2, 4, 7, 8, 9, 10, 11, 12, 13, 14, 15, …
#> $ completions                 <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ attempts                    <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ passing_yards               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ passing_tds                 <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ interceptions               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ sacks                       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ sack_fumbles_lost           <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ passing_air_yards           <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ passing_yards_after_catch   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ passing_first_downs         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ passing_epa                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ passing_2pt_conversions     <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ dakota                      <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ carries                     <int> 16, 9, 3, 6, 13, 9, 17, 7, 15, 10, 5, 13, …
#> $ rushing_yards               <dbl> 60, 33, 2, 27, 39, 23, 54, 11, 35, 29, 10,…
#> $ rushing_tds                 <int> 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ rushing_fumbles_lost        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ rushing_first_downs         <dbl> 4, 1, 0, 0, 2, 1, 4, 0, 1, 1, 0, 2, 4, 0, …
#> $ rushing_epa                 <dbl> 6.24877114, -1.43495017, -1.53995173, 0.21…
#> $ rushing_2pt_conversions     <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ receptions                  <int> 1, 3, 0, 2, 0, 1, 1, 2, 1, 1, 2, 1, 2, 0, …
#> $ targets                     <int> 1, 4, 1, 2, 0, 2, 1, 3, 1, 1, 2, 1, 3, 0, …
#> $ receiving_yards             <dbl> 7, 18, 0, 8, 0, 2, 7, 2, 1, 21, 7, 4, 7, 0…
#> $ receiving_tds               <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, …
#> $ receiving_fumbles_lost      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ receiving_air_yards         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ receiving_yards_after_catch <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ receiving_first_downs       <dbl> 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 1, 0, 0, 0, …
#> $ receiving_epa               <dbl> 0.29237815, 0.37700888, -0.69957773, -0.22…
#> $ receiving_2pt_conversions   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ special_teams_tds           <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ fantasy_points              <dbl> 12.7, 5.1, 0.2, 3.5, 3.9, 2.5, 6.1, 1.3, 3…
#> $ fantasy_points_ppr          <dbl> 13.7, 8.1, 0.2, 5.5, 3.9, 3.5, 7.1, 3.3, 4…

# }