Get a Situation Report on System, nflverse Package Versions and Dependencies

This function gives a quick overview of the versions of R and the operating system as well as the versions of nflverse packages, options, and their dependencies. It's primarily designed to help you get a quick idea of what's going on when you're helping someone else debug a problem.

Usage

report(...)

Arguments

...

Arguments passed on to nflreadr::nflverse_sitrep

pkg

a character vector naming installed packages, or NULL (the default) meaning all nflverse packages. The function checks internally if all packages are installed and informs if that is not the case.

recursive

a logical indicating whether dependencies of pkg and their dependencies (and so on) should be included. Can also be a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"). Character string "all" is shorthand for that vector, character string "most" for the same vector without "Enhances", character string "strong" (default) for the first three elements of that vector.

redact_path

a logical indicating whether options that contain "path" in the name should be redacted, default = TRUE

Details

See nflreadr::nflverse_sitrep for details.

Examples

# \donttest{
report(recursive = FALSE)
#> ℹ You've asked for the packages nfl4th and nflverse which are not installed. 
#> They are skipped.
#> ── System Info ─────────────────────────────────────────────────────────────────
#> • R version 4.2.1 (2022-06-23)   • Running under: Ubuntu 20.04.5 LTS
#> ── nflverse Packages ───────────────────────────────────────────────────────────
#> • nflreadr (1.3.0.05)    • nflseedR (1.1.0)    
#> • nflfastR (4.4.0.9010)  • nflplotR (1.1.0)    
#> ── nflverse Options ────────────────────────────────────────────────────────────
#> No options set for nflreadr, nflfastR, nflseedR, and nflplotR
nflverse_sitrep(pkg = "nflreadr", recursive = TRUE)
#> ── System Info ─────────────────────────────────────────────────────────────────
#> • R version 4.2.1 (2022-06-23)   • Running under: Ubuntu 20.04.5 LTS
#> ── nflverse Packages ───────────────────────────────────────────────────────────
#> • nflreadr (1.3.0.05)
#> ── nflverse Options ────────────────────────────────────────────────────────────
#> No options set for nflreadr
#> ── nflverse Dependencies ───────────────────────────────────────────────────────
#> • cachem (1.0.6)  • data.table (1.14.2)  • memoise  (2.0.1)  
#> • cli    (3.4.0)  • fastmap    (1.1.0)   • rappdirs (0.3.3)  
#> • curl   (4.3.2)  • glue       (1.6.2)   • rlang    (1.0.5)  
#> ────────────────────────────────────────────────────────────────────────────────
# }