NFL Team names, colors and logo urls.

NFL Team names, colors and logo urls.

teams_colors_logos

Format

A data frame with 36 rows and 10 variables containing NFL team level information, including franchises in multiple cities:

team_abbr

Team abbreviation

team_name

Complete Team name

team_id

Team id used in the roster function

team_nick

Nickname

team_color

Primary color

team_color2

Secondary color

team_color3

Tertiary color

team_color4

Quaternary color

team_logo_wikipedia

Url to Team logo on wikipedia

team_logo_espn

Url to higher quality logo on espn

team_wordmark

Url to team wordmarks

The colors are taken from Lee Sharpe's teamcolors.csv who has taken them from the teamcolors package created by Ben Baumer and Gregory Matthews. The Wikipeadia logo urls are taken from Lee Sharpe's logos.csv Team wordmarks from nfl.com

Examples

# \donttest{
teams_colors_logos

#> # A tibble: 36 x 11
#>    team_abbr team_name team_id team_nick team_color team_color2 team_color3
#>    <chr>     <chr>     <chr>   <chr>     <chr>      <chr>       <chr>      
#>  1 ARI       Arizona … 3800    Cardinals #97233f    #000000     #ffb612    
#>  2 ATL       Atlanta … 0200    Falcons   #a71930    #000000     #a5acaf    
#>  3 BAL       Baltimor… 0325    Ravens    #241773    #000000     #9e7c0c    
#>  4 BUF       Buffalo … 0610    Bills     #00338d    #c60c30     #0c2e82    
#>  5 CAR       Carolina… 0750    Panthers  #0085ca    #000000     #bfc0bf    
#>  6 CHI       Chicago … 0810    Bears     #0b162a    #c83803     #0b162a    
#>  7 CIN       Cincinna… 0920    Bengals   #000000    #fb4f14     #000000    
#>  8 CLE       Clevelan… 1050    Browns    #fb4f14    #22150c     #a5acaf    
#>  9 DAL       Dallas C… 1200    Cowboys   #002244    #b0b7bc     #acc0c6    
#> 10 DEN       Denver B… 1400    Broncos   #002244    #fb4f14     #00234c    
#> # … with 26 more rows, and 4 more variables: team_color4 <chr>,
#> #   team_logo_wikipedia <chr>, team_logo_espn <chr>, team_wordmark <glue>

# }